feat: 통합된 MeetingDetailCard 컴포넌트로 리팩토링
- MeetingDetailHeader와 MeetingActions를 MeetingDetailCard로 통합
- Card 컴포넌트 구조를 활용하여 일관된 스타일링 적용
- 주최자 크라운 아이콘 추가 (crown.svg 사용)
- Chip 컴포넌트를 사용한 태그 시스템 적용
- 불필요한 스토리 파일들 제거
- 이슈 #33 관련
